Move type to separate file
