+++ src/input.spec.ts
@@ -1,6 +1,7 @@
 import yaml from "js-yaml";
 import fs from "fs";
-import { NetlifyCMSConfig, loadConfig } from "./input";
+import { loadConfig } from "./input";
+import type { NetlifyCMSConfig } from "./types";
 
 const mockConfigObject = yaml.load(
   fs.readFileSync("kitchen-sink.yml", "utf-8"),
