+++ src/types/config.ts
@@ -1,0 +1,5 @@
+import { Collection } from "./fields";
+
+export interface NetlifyCMSConfig {
+  collections?: Collection[];
+}
