+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export type { Field, Collection } from "./fields";
 export type { Widget } from "./widgets";
 export type { NetlifyTsOptions } from "./options";
+export type { NetlifyCMSConfig } from "./config";
